[indirect-executor] move trait definitions to separate module.
